+++ src/cinema6/cinema6.js
@@ -181,6 +181,10 @@ function( angular , eventsEmitter     ,    postmessagePostmessage  ) {
                         return cache.put(model._type + ':' + model.id, model);
                     }
 
+                    function cleanup() {
+                        delete self._pending;
+                    }
+
                     if (this._erased) {
                         return $q.reject('Cannot save an erased record.');
                     }
@@ -192,7 +196,8 @@ function( angular , eventsEmitter     ,    postmessagePostmessage  ) {
                         (this._pending = getAdapter()[this.id ?
                             'update' : 'create'](this._type, this.pojoify())
                             .then(update)
-                            .then(cacheModel));
+                            .then(cacheModel)
+                            .finally(cleanup));
                 },
                 erase: function() {
                     var self = this;